v0.0.01d.20160516 - Alpha test version - Fixed the time triggers not being able to fire due to taskId parameter missing

+++ smartapps/ady624/core.src/core.groovy
@@ -1749,7 +1749,7 @@ private getNextConditionId() {
     while (settings.findAll { it.key == "condParent" + nextId }) {
     	nextId++
     }
-    return nextId
+    return (int) nextId
 }
 
 //helper function for getNextId
@@ -1787,7 +1787,7 @@ private getNextActionId() {
     while (settings.findAll { it.key == "actParent" + nextId }) {
     	nextId++
     }
-    return nextId
+    return (int) nextId
 }
 
 private updateAction(action) {
